move SentryTracerFinishStatus to own file
