+++ dart/lib/src/sentry_tracer_finish_status.dart
@@ -1,0 +1,17 @@
+import 'package:meta/meta.dart';
+
+import '../sentry.dart';
+
+@internal
+class SentryTracerFinishStatus {
+  final bool finishing;
+  final SpanStatus? status;
+
+  SentryTracerFinishStatus.finishing(SpanStatus? status)
+      : finishing = true,
+        status = status;
+
+  SentryTracerFinishStatus.notFinishing()
+      : finishing = false,
+        status = null;
+}
